# 8285398: Cache the results of constraint checks