@djelinski 2022-04-21
before:
```
Benchmark                 (resume)  (tlsVersion)   Mode  Cnt     Score      Error  Units
SSLHandshake.doHandshake      true       TLSv1.2  thrpt    5  2165.081 ± 440.204  ops/s
SSLHandshake.doHandshake      true           TLS  thrpt    5   534.681 ± 146.931  ops/s
SSLHandshake.doHandshake     false       TLSv1.2  thrpt    5   369.104 ±  11.143  ops/s
SSLHandshake.doHandshake     false           TLS  thrpt    5   253.903 ±  58.056  ops/s
```
after:
```
Benchmark                 (resume)  (tlsVersion)   Mode  Cnt      Score     Error  Units
SSLHandshake.doHandshake      true       TLSv1.2  thrpt    5  10440.501 ± 478.177  ops/s
SSLHandshake.doHandshake      true           TLS  thrpt    5    762.995 ±  33.842  ops/s
SSLHandshake.doHandshake     false       TLSv1.2  thrpt    5    440.471 ±  52.867  ops/s
SSLHandshake.doHandshake     false           TLS  thrpt    5    305.928 ±  57.847  ops/s
```
After this patch the HandshakeContext initialization practically disappears from the CPU profile; it only takes ~5% in TLS1.2 resumption, and much less in the remaining scenarios.